fix multilinear poly compilation add mul
