+++ math/src/fft/polynomial.rs
@@ -5,7 +5,7 @@ use crate::{
         element::FieldElement,
         traits::{IsFFTField, RootsConfig},
     },
-    polynomial::polynomial::Polynomial,
+    polynomial::Polynomial,
 };
 
 #[cfg(feature = "cuda")]
